Don't use invalid self-closing tag in example
Self-closing angular js directives like this one are considered invalid. This use case happens to work in many situations, but it can also have undesired effects. For example, if you use `<react-component />` as the sibling of another tag, it can "swallow up" the sibling, causing it to never be rendered.